Unnecessarily large commit.
Made current logged in user a variable in GameData, so need to pass it
between screens.

Got the weekly pick's name from dropdown box. Still needs to be written to
file.

Added in a function to make it easier to add contestants when reading in
the file for the first time. Client cannot add contestants, so the new
function should not destroy the backend.